+++ src/main.cpp
@@ -31,8 +31,13 @@ int main() {
     // Power Manage Init
     pm_init();
 
-    // LED
+    // Judge whether board is Pico or Waveshare RP2040-LCD-0.96
+    // PICO_DEFAULT_LED_PIN: GPIO25
+    //    Raspberry Pi Pico: board LED (connected to LCD only)
+    //    Waveshare RP2040-LCD-0.96: LCD BLK (pulled-up to 3.3K and connected to driver of backlight)
     gpio_init(PICO_DEFAULT_LED_PIN);
+    gpio_set_dir(PICO_DEFAULT_LED_PIN, GPIO_IN);
+    board_type_t board_type = gpio_get(PICO_DEFAULT_LED_PIN) ? WAVESHARE_RP2040_LCD_096 : RASPBERRY_PI_PICO;
     gpio_set_dir(PICO_DEFAULT_LED_PIN, GPIO_OUT);
     gpio_put(PICO_DEFAULT_LED_PIN, 0);
 
@@ -42,12 +47,23 @@ int main() {
         sleep_ms(25);
     }
     printf("Raspberry Pi Pico Player ver. %s\n", VersionStr);
+    switch (board_type) {
+        case RASPBERRY_PI_PICO:
+            printf("Board: Raspberry Pi Pico\n");
+            break;
+        case WAVESHARE_RP2040_LCD_096:
+            printf("Board: Waveshare RP2040-LCD-0.96\n");
+            break;
+        default:
+            printf("Board: unknown\n");
+            break;
+    }
 
     // ADC Initialize
     adc_init();
 
     // UI initialize
-    ui_init();
+    ui_init(board_type);
 
     // UI Loop (infinite)
     const int LoopCycleMs = UIMode::UpdateCycleMs; // loop cycle (50 ms)